requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
aiosqlite==0.19.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
httpx==0.25.2
